requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.0
speechrecognition==3.10.0
pydub==0.25.1
nltk==3.8.1
pymongo==4.6.1
python-jose==3.3.0
validators==0.22.0 